SEP6-130: Highest Revenue by year

+++ src/MovieManagement.TMDB.Api/Dtos/MovieDto.cs
@@ -4,7 +4,7 @@ public class MovieDto
 {
     public int Id { get; set; }
     public int Budget { get; set; }
-    public int Revenue { get; set; }
+    public float Revenue { get; set; }
     public string Title { get; set; } = default!;
     [JsonPropertyName("release_date")] public string ReleaseDate { get; set; } = default!;
     [JsonPropertyName("vote_average")] public double VoteAverage { get; set; }
